Move signature endpoints to plugin
